+++ src/app/auth/verify-user-by-email/authenticate/page.tsx
@@ -11,7 +11,7 @@ import { HttpStatus } from '../../../../server-action/backend/status';
 import { OneTimeTokenPurposeValues } from '../../../../server-action/backend/auth/type';
 import ErrorModal from '../../../../components/ErrorModal';
 import GuideModal from '../../../../components/GuideModal';
-import { AUTH_LOGIC_ROUTE } from '../../route';
+import { AUTH_LOGIC_ROUTE } from '../../../../route/auth/route';
 
 interface AuthenticateState {
   errorMessage : string;
