+++ src/server-action/backend/cookie.ts
@@ -3,7 +3,7 @@ import { cookies } from 'next/headers';
 import { redirect } from 'next/navigation';
 import { OneTimeToken, SignInResponse } from './auth/type';
 import { ResponseCookies } from 'next/dist/compiled/@edge-runtime/cookies';
-import { AUTH_LOGIC_ROUTE } from '../../app/auth/route';
+import { AUTH_LOGIC_ROUTE } from '../../route/auth/route';
 const ENUM_COOKIE_KEY = {
     SIGN_IN_RESPONSE: 'SignInResponse',
     ONE_TIME_TOKEN: 'OneTimeToken',
